ovs-dpdk: change dpdk-driver default to None
Some NICs do not work with vfi-pci or uio_pci_generic. E.g. mxl4/mlx5
which uses relies on the OFED or Kernel drivers (post 4.14).
In this cases we don't want to generate entries in /etc/dpdk/inerfaces.

Here we change the configuration processing behavior. The charm will omit
adding entries in the aforementioned file when the value is not set.

The default value is changed to empty (i.e. None)

Change-Id: I2fb9f0404adbbee0f298729467794e172bae2d98